Crate gamepad_web

Event-based Gamepad Web Backend

GitHub repository

This crate extends the Gamepad Web API, to emit events for button and axis state changes.

Our goal is to resemble a native API, which emits an event for each state change, unlike the Web API, which only allows state querying.

Quick Start

use gamepad_web::*;

// start listening for Gamepad events
let mut monitor = Monitor::new();

fn update() {
    // process new input events
    while let Some(event) = monitor.poll_mapped() {
        match event {
            MappedEvent::ButtonPress(Button::South) => // "A" on Xbox
                player.jump(),
            MappedEvent::Axis(Axis::LeftStickX, x) =>
                player.set_velocity(x),
            _ => ()
        }
    }
}

Structs

Event
An event, including connections, disconnections, and button/axis input.

GamepadDescription
Information about a gamepad. Remains constant while the gamepad is connected.

GamepadState
Snapshot of a gamepad's state at one instant.

Monitor
Listen for gamepad events

Enums

Axis
A named axis on the standard gamepad.

Button
A named button on the standard gamepad.

EventData
The data associated with an event, not including the gamepad itself.

GamepadMappingType
The set of known gamepad layout mappings.

GamepadStateChange
A change in axis or button state.

MappedEvent
An input event, mapped to one of the standard buttons or axes.

Traits

Mapping
A relation between indices and names, for buttons and axes.