Struct gamepad_web::GamepadState

pub struct GamepadState { /* fields omitted */ }

Snapshot of a gamepad's state at one instant.

Methods

impl GamepadState

pub fn timestamp(&self) -> f64

The DOMHighRestTimeStamp when this gamepad was last updated.

pub fn axis_count(&self) -> usize

The number of axes this gamepad reports.

pub fn button_count(&self) -> usize

The number of buttons this gamepad reports.

pub fn axis(&self, index: usize) -> f64

Get the value of an axis [-1–1].

Panics

Panics if index >= axis_count().

pub fn button_pressed(&self, index: usize) -> bool

Is the button at this index pressed / held down?

Panics

Panics if index >= button_count().

pub fn button_value(&self, index: usize) -> f64

How much is the button at this index pressed / held down?

For most buttons on most devices, this is either 0.0 (not pressed) or 1.0 (pressed). The main exceptions are triggers, which smoothly transition from 0.0 to 1.0 when pressed.

Panics

Panics if index >= button_count().

pub fn changes_since<'a>(
&'a self,
previous: &'a Self
) -> impl Iterator<Item = GamepadStateChange> + 'a

Find the changes in this state, compared to a previous state.

The timestamps are not considered. previous could be newer than self, this still returns changes from previous to self.